Create cubeOutput.txt
This is a stream of data lines created by the cube .clj  file . Copy and paste the .clj file into a clojure REPL to create this file.     
It creates a red box on lowest plane and a blue box on highest plane. 

Generated on latitude laptop running Xubuntu, Clojure 1.6.   The terminal program used was GTKterm 115200-8-N-1. Previously Hyperterm on WIN XP was used. The Arduino is recognised as a Leonardo.